Completed checking for empty string and unique names when creating exercises
Created checkIfExerciseAlreadyExists()

When populateButtonsForNewRoutines() pushes an exercise name into exisitingExerciseArray, it changes it to lowercase so all the elements in that array is standardized to lowercasee

+++ scripts/scripts.js
@@ -1,10 +1,18 @@
 /**
- * Global variables
+ * * Global Variables
  */
-const alreadyCreatedButtonsArray = []; // Used for populateButtonsForNewRoutines()
+const existingExerciseArray = [];
 
 /**
- * Runs this function upon page load
+ * * Helper Functions
+ */
+
+function checkIfExerciseAlreadyExists(exercise) {
+     return existingExerciseArray.includes(exercise.toLowerCase());
+}
+
+/**
+ * * On page load, run this function
  */
 window.onload = pageLoad();
 
@@ -33,7 +41,6 @@ function toggleViewOnAndOff(divId) {
 /**
  * This function handles the saving of new exercises into localStorage
  */
-// TODO: Make it so that exercise names must be unique
 function saveNewExerciseToLocalStorage() {
      event.preventDefault();
 
@@ -51,23 +58,51 @@ function saveNewExerciseToLocalStorage() {
      // Store user input value
      const exerciseName = document.getElementById("new-exercise-input").value;
 
-     // Create a new Exercise Object, its name will be the value from input
-     const newExercise = new Exercise(exerciseName, "");
+     // Only create a new exercise if it isn't an empty string and it doesn't already exist
+     if (exerciseName.trim().length > 0 && checkIfExerciseAlreadyExists(exerciseName) === false) {
+          // Create a new Exercise Object, its name will be the value from input
+          const newExercise = new Exercise(exerciseName, "");
 
-     // Pushes the new Exercise Object into the existing exercise data
-     existingExerciseData.push(newExercise);
+          // Pushes the new Exercise Object into the existing exercise data
+          existingExerciseData.push(newExercise);
 
-     // Save existing data with the new exercise to localStorage
-     localStorage.setItem("exerciseDataKey", JSON.stringify(existingExerciseData));
+          // Save existing data with the new exercise to localStorage
+          localStorage.setItem("exerciseDataKey", JSON.stringify(existingExerciseData));
 
-     // Resets the input field to be empty
+          // Populates routine section with a new button for the new exercise
+          populateButtonsForNewRoutines();
+     }
+
+     // Always resets the input field to be empty
      document.getElementById("new-exercise-input").value = "";
-
-     // Populates routine section with a new button for the new exercise
-     populateButtonsForNewRoutines();
 
      // ! for testing area, remove later
      testDisplayExercises();
+}
+
+/**
+ * This function populates the exercise buttons that a user can click on to add exercises to a new routine
+ */
+
+function populateButtonsForNewRoutines() {
+     // Grabs existing exercises data from localStorage and converts it into an object
+     const existingExerciseDataInStringForm = localStorage.getItem("exerciseDataKey");
+     const existingExerciseDataInObjectForm = existingExerciseDataInStringForm ? JSON.parse(existingExerciseDataInStringForm) : [];
+
+     for (i = 0; i < existingExerciseDataInObjectForm.length; i++) {
+          // If an exercise already exists as a button, don't create another one for it
+          if (checkIfExerciseAlreadyExists(existingExerciseDataInObjectForm[i].name)) {
+               continue;
+          } else {
+               const newButton = document.createElement("button");
+               const exerciseName = existingExerciseDataInObjectForm[i].name;
+               // TODO: Should add a class/ids to these buttons?
+
+               newButton.innerHTML = exerciseName;
+               document.getElementById("div-new-routine").append(newButton);
+               existingExerciseArray.push(exerciseName.toLowerCase());
+          }
+     }
 }
 
 // * WORKING ON CODE BELOW
@@ -78,25 +113,4 @@ function testDisplayExercises() {
      document.getElementById("testing-area").innerHTML = exercises;
 }
 
-function populateButtonsForNewRoutines() {
-     // Grabs existing exercises data from localStorage and converts it into an object
-     const existingExerciseDataInStringForm = localStorage.getItem("exerciseDataKey");
-     const existingExerciseDataInObjectForm = existingExerciseDataInStringForm ? JSON.parse(existingExerciseDataInStringForm) : [];
-
-     for (i = 0; i < existingExerciseDataInObjectForm.length; i++) {
-          let doesExerciseAlreadyExist = alreadyCreatedButtonsArray.includes(existingExerciseDataInObjectForm[i].name);
-
-          // If an exercise already exists as a button, don't create another one for it
-          if (doesExerciseAlreadyExist) {
-               continue;
-          } else {
-               const newButton = document.createElement("button");
-               const exerciseName = existingExerciseDataInObjectForm[i].name;
-               // TODO: Should add a class/ids to these buttons?
-
-               newButton.innerHTML = exerciseName;
-               document.getElementById("div-new-routine").append(newButton);
-               alreadyCreatedButtonsArray.push(exerciseName);
-          }
-     }
-}
+console.log(existingExerciseArray);
